// src/typings/Pokemon.ts
export type PokemonBase = {
  id: string;
  url: string;
  name: string;
  image: string;
};

export type PokemonDetail = PokemonBase & {
  id: string;
  name: string;
  height: number;
  weight: number;
  abilities: Array<{
    ability: {
      name: string;
    };
  }>;
  types: Array<{ type: { name: string } }>;
  stats: Array<{ name: string; value: number }>;
  moves: Array<{ move: { name: string } }>;
  species: string;
  sprites: {
    front_default: string;
    other: {
      "official-artwork": {
        front_default: string;
      };
    };
  };
};